backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
sqlmodel==0.0.24
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.16.5
python-dotenv==1.1.1
pydantic==2.10.6
pydantic-settings==2.10.1
PyJWT
bcrypt==4.1.2
email-validator==2.1.0
pdfplumber==0.11.4
python-multipart==0.0.12

